tests/wpt/web-platform-tests/css/css-syntax/charset/page-windows-1251-css-utf8-bom.html

<!doctype html>
<title>CSS charset: page windows-1251, CSS UTF-8 BOM</title>
<link rel=help href="https://drafts.csswg.org/css-syntax-3/#determine-the-fallback-encoding">
<meta charset=windows-1251>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel=stylesheet href="support/utf8-bom.css">
<div id=log></div>
<div id=&#xfffd;></div>
<script>
var elm = document.getElementById('\ufffd');
var t = async_test();
onload = t.step_func(function(){
  assert_equals(getComputedStyle(elm, '').visibility, 'hidden');
  this.done();
});
</script>
